sqlite_twitterprofiles_to_RAG/requirements.txt
aiohttp==3.9.1
aiosqlite==0.19.0
pandas==2.0.3
requests==2.31.0
python-dotenv==1.1.0
